requirements.txt
pandas>=2.0.3
openpyxl>=3.1.2
python-dotenv>=1.0.0
schedule>=1.2.0
flask>=2.3.0
requests>=2.31.0
oracledb>=3.0.0 